pyproject.toml
[project]
name = "easymaas"
version = "0.1.0"
description = "Easily wrap Python code as OpenAI-compatible services"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "llipa"}
]
dependencies = [
    "fastapi>=0.115.12",
    "psutil>=7.0.0",
    "pyyaml>=6.0.2",
    "uvicorn>=0.34.1",
    "click>=8.1.7",
    "pydantic>=2.0.0",
]

[project.scripts]
easymaas = "easymaas.cli.commands:cli"

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"
